@@ -1,20 +1,18 @@
 name: wasm_interop
-version: 1.0.0-dev.1.0
-description: Dart wrapper for WebAssembly JavaScript API
+version: 2.0.0
+description: This package provides a Dart wrapper for WebAssembly JavaScript API
 author: Alexey Knyazev
 homepage: https://github.com/lexaknyazev/wasm.dart
-documentation:
 
 environment:
-  sdk: '>=2.1.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  js: ^0.6.1
-  meta: ^1.1.5
+  js: ^0.6.3
+  meta: ^1.4.0
 
 dev_dependencies:
-  build_runner: ^1.2.0
+  build_runner: any
   build_test: any
-  build_web_compilers: ^1.0.0
-  build_node_compilers: ^0.2.1
-  test: ^1.5.2
+  build_web_compilers: any
+  test: ^1.17.5
